+++ src/gguf/fileReaders/GgufFsFileReader.ts
@@ -2,7 +2,7 @@ import fs from "node:fs/promises";
 import retry from "async-retry";
 import {withLock} from "lifecycle-utils";
 import {GgufReadOffset} from "../utils/GgufReadOffset.js";
-import {defaultExtraAllocationSize, ggufDefaultRetryOptions} from "../../consts.js";
+import {defaultExtraAllocationSize, ggufDefaultRetryOptions} from "../consts.js";
 import {GgufFileReader} from "./GgufFileReader.js";
 
 type GgufFsFileReaderOptions = {
