+++ src/gguf/parser/parseGguf.ts
@@ -1,10 +1,10 @@
 import {InvalidGgufMagicError} from "../errors/InvalidGgufMagicError.js";
 import {getConsoleLogPrefix} from "../../utils/getConsoleLogPrefix.js";
 import {UnsupportedError} from "../../utils/UnsupportedError.js";
-import {GgufReadOffset} from "./utils/GgufReadOffset.js";
-import {GgufFileReader, valueTypeToBytesToRead} from "./fileReaders/GgufFileReader.js";
-import {GgufFileInfo, GgufVersionParserOptions, GgufVersionParserResult} from "./types/GgufFileInfoTypes.js";
-import {GgufV2Parser} from "./parsers/GgufV2Parser.js";
+import {GgufReadOffset} from "../utils/GgufReadOffset.js";
+import {GgufFileReader, valueTypeToBytesToRead} from "../fileReaders/GgufFileReader.js";
+import {GgufFileInfo, GgufVersionParserOptions, GgufVersionParserResult} from "../types/GgufFileInfoTypes.js";
+import {GgufV2Parser} from "./GgufV2Parser.js";
 
 const ggufMagic = "GGUF";
