+++ src/gguf/types/GgufMetadataTypes.ts
