+++ src/gguf/utils/GgufReadOffset.ts
